# APIDocumentation/openapi.yaml
openapi: 3.1.0
info:
  title: Minowa.ai UserApp API
  version: "0.2.0"
  description: |
    OpenAPI contract for the patient-facing UserApp API.

    **Scope.** The contract covers the auth/session surface, the
    stable health-data endpoints (health-inputs, blood-pressure, temperature,
    weight, observations), and account-key management. It does NOT cover
    push/pull sync endpoints (dietary settings),
    documents/fax, or wearable sync (Garmin, HealthKit).
    Those are intentionally deferred until their schemas stabilize.

    **Doctrine.** This file is the contract. If the implementation diverges
    from what is declared here, either the implementation or the spec is
    wrong — schemathesis will catch the drift in CI.

    **v1 / v2 aliasing.** Every `/api/v1/*` path declared below is also
    served at the corresponding `/api/v2/*` path with identical request and
    response shapes (the v2 routes are decorator aliases on the same handler).
    Only v1 paths are declared here to keep the spec tight; the v2 paths
    follow the same contract.

    **To add an endpoint.** Declare the path, methods, request body, and
    response schemas. Run `schemathesis run openapi.yaml --dry-run` locally
    to lint, then enable the live-fuzz job once the endpoint is accurate.
servers:
  - url: http://localhost:80
    description: Home Edition appliance (local / LAN)
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/blood-pressure:
    get:
      operationId: listBloodPressure
      tags: [vitals]
      summary: List blood pressure readings (paginated, optional date range)
      parameters:
        - $ref: "#/components/parameters/startDate"
        - $ref: "#/components/parameters/endDate"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: Paginated readings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BloodPressureList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      operationId: logBloodPressure
      tags: [vitals]
      summary: Log a blood pressure reading
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BloodPressureCreate"
      responses:
        "201":
          description: Reading logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/blood-pressure/{reading_id}:
    delete:
      operationId: deleteBloodPressure
      tags: [vitals]
      summary: Delete a blood pressure reading
      parameters:
        - $ref: "#/components/parameters/readingId"
      responses:
        "200":
          description: Reading deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ============================================================================
  # VITALS — TEMPERATURE
  # ============================================================================
  /api/v1/temperature:
    get:
      operationId: listTemperature
      tags: [vitals]
      summary: List temperature readings (paginated, optional date range)
      parameters:
        - $ref: "#/components/parameters/startDate"
        - $ref: "#/components/parameters/endDate"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: Paginated readings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemperatureList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      operationId: logTemperature
      tags: [vitals]
      summary: Log a temperature reading
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemperatureCreate"
      responses:
        "201":
          description: Reading logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ============================================================================
  # VITALS — WEIGHT
  # ============================================================================
  /api/v1/weight:
    get:
      operationId: listWeight
      tags: [vitals]
      summary: List weight readings (paginated, optional date range)
      parameters:
        - $ref: "#/components/parameters/startDate"
        - $ref: "#/components/parameters/endDate"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: Paginated readings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeightList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      operationId: logWeight
      tags: [vitals]
      summary: Log a weight reading
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WeightCreate"
      responses:
        "201":
          description: Reading logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/weight/{weight_id}:
    delete:
      operationId: deleteWeight
      tags: [vitals]
      summary: Delete a weight reading
      parameters:
        - $ref: "#/components/parameters/weightId"
      responses:
        "200":
          description: Reading deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/observations:
    get:
      operationId: listObservations
      tags: [vitals]
      summary: List free-text observations (paginated)
      parameters:
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: Paginated observations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ObservationList"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      operationId: createObservation
      tags: [vitals]
      summary: Create a free-text observation (server-side embedded in background)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ObservationCreate"
      responses:
        "201":
          description: Observation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedId"
          links:
            update:
              operationId: updateObservation
              parameters:
                obs_id: "$response.body#/id"
            patch:
              operationId: patchObservation
              parameters:
                obs_id: "$response.body#/id"
            delete:
              operationId: deleteObservation
              parameters:
                obs_id: "$response.body#/id"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/observations/{obs_id}:
    put:
      operationId: updateObservation
      tags: [vitals]
      summary: Replace an observation's text and timestamp (re-embeds in background)
      parameters:
        - $ref: "#/components/parameters/obsId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ObservationCreate"
      responses:
        "200":
          description: Observation updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

    patch:
      operationId: patchObservation
      tags: [vitals]
      summary: Toggle the mental-health flag on an observation
      parameters:
        - $ref: "#/components/parameters/obsId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mental_health_flag]
              properties:
                mental_health_flag:
                  type: boolean
      responses:
        "200":
          description: Observation updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"

    delete:
      operationId: deleteObservation
      tags: [vitals]
      summary: Delete an observation
      parameters:
        - $ref: "#/components/parameters/obsId"
      responses:
        "200":
          description: Observation deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Session token returned by `/api/v1/login`, or a long-lived API key
        created via `/api/v1/api-keys`. Both are accepted on the same header.
  parameters:
    limit:
      name: limit
      in: query
      required: false
      description: Page size (clamped server-side; default varies by endpoint).
      schema:
        type: integer
        minimum: 1
    offset:
      name: offset
      in: query
      required: false
      description: Page offset (0-based).
      schema:
        type: integer
        minimum: 0
    startDate:
      name: start_date
      in: query
      required: false
      description: Inclusive lower bound on the measurement date (YYYY-MM-DD).
      schema:
        type: string
        format: date
    endDate:
      name: end_date
      in: query
      required: false
      description: Inclusive upper bound on the measurement date (YYYY-MM-DD).
      schema:
        type: string
        format: date
...
    readingId:
      name: reading_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    weightId:
      name: weight_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    obsId:
      name: obs_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    MessageEnvelope:
      type: object
      required: [message]
      properties:
        message:
          type: string
        success:
          type: boolean
    CreatedId:
      type: object
      required: [id, message]
      properties:
        id:
          type: string
          format: uuid
        message:
          type: string
    Pagination:
      type: object
      required: [total, limit, offset, has_more]
      properties:
        total:
          type: integer
          minimum: 0
        limit:
          type: integer
          minimum: 1
        offset:
          type: integer
          minimum: 0
        has_more:
          type: boolean

    # --- Auth --------------------------------------------------------------
...
    BloodPressureReading:
      type: object
      required: [id, systolic, diastolic, timestamp]
      properties:
        id:
          type: string
          format: uuid
        systolic:
          type: integer
          minimum: 40
          maximum: 300
        diastolic:
          type: integer
          minimum: 20
          maximum: 200
        pulse:
          type:
            - integer
            - "null"
          minimum: 20
          maximum: 250
        timestamp:
          type: string
          format: date-time
          description: |
            UTC ISO-8601 timestamp. Backed by the DB column `measured_at`
            (renamed on the wire for consistency with other vitals endpoints).
    BloodPressureCreate:
      type: object
      required: [timestamp, systolic, diastolic]
      properties:
        timestamp:
          type: string
          description: |
            Local datetime string in the user's home timezone. Server
            converts to UTC before storing.
        systolic:
          type: integer
          minimum: 40
          maximum: 300
        diastolic:
          type: integer
          minimum: 20
          maximum: 200
        heart_rate:
          type:
            - integer
            - "null"
          minimum: 20
          maximum: 250
          description: |
            Pulse at the time of the reading. Stored as `pulse` in the DB
            and read back as `pulse` on `GET /blood-pressure` — the write
            side calls it `heart_rate` for legacy reasons.
    BloodPressureList:
      type: object
      required: [entries, pagination]
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/BloodPressureReading"
        pagination:
          $ref: "#/components/schemas/Pagination"

    # --- Vitals: temperature ----------------------------------------------
...
    TemperatureCreate:
      type: object
      required: [timestamp, temperature]
      properties:
        timestamp:
          type: string
          description: Local datetime string; server converts to UTC.
        temperature:
          type: number
        unit:
          type: string
          description: Defaults to `F` if omitted.
    TemperatureList:
      type: object
      required: [entries, pagination]
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/TemperatureReading"
        pagination:
          $ref: "#/components/schemas/Pagination"

    # --- Vitals: weight ----------------------------------------------------
...
    WeightCreate:
      type: object
      required: [timestamp, weight]
      properties:
        timestamp:
          type: string
          description: Local datetime string; server converts to UTC.
        weight:
          type: number
        unit:
          type: string
          description: Defaults to `lbs` if omitted.
    WeightList:
      type: object
      required: [entries, pagination]
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/WeightReading"
        pagination:
          $ref: "#/components/schemas/Pagination"

    # --- Vitals: observations ---------------------------------------------
...
    ObservationCreate:
      type: object
      required: [observation]
      properties:
        observation:
          type: string
        timestamp:
          type: string
          description: |
            Local datetime string; server converts to UTC. Defaults to "now"
            when omitted.
        source_type:
          type: string
          description: Defaults to `text` if omitted.
        mental_health_flag:
          type: boolean
    ObservationList:
      type: object
      required: [entries, pagination]
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/Observation"
        pagination:
          $ref: "#/components/schemas/Pagination"
  responses:
    BadRequest:
      description: Invalid request body or query parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Target resource does not exist or is not owned by the caller
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...